[helm nats 1.x] add tlsCA option
