+++ helm/charts/nats/files/nats-box/deployment/container.yaml
@@ -26,12 +26,17 @@ command:
 args:
 - trap true INT TERM; sleep infinity & wait
 volumeMounts:
+# contexts secret
 - name: contexts
   mountPath: /etc/nats-contexts
+# contents secret
 {{- if .hasContentsSecret }}
 - name: contents
   mountPath: /etc/nats-contents
 {{- end }}
+# tlsCA
+{{- include "nats.tlsCAVolumeMount" $ }}
+# secrets
 {{- range (include "natsBox.secretNames" $ | fromJson).secretNames }}
 - name: {{ .name | quote }}
   mountPath: {{ .dir | quote }}
