+++ helm/charts/nats/files/nats-box/deployment/pod-template.yaml
@@ -27,6 +27,8 @@ spec:
     secret:
       secretName: {{ .Values.natsBox.contentsSecret.name }}
   {{- end }}
+  # tlsCA
+  {{- include "nats.tlsCAVolume" $ | nindent 2 }}
   # secrets
   {{- range (include "natsBox.secretNames" $ | fromJson).secretNames }}
   - name: {{ .name | quote }}
