+++ helm/charts/nats/files/service.yaml
@@ -16,8 +16,6 @@ spec:
   {{- $tlsEnabled := false }}
   {{- if hasKey $configProtocol "tls" }}
     {{- $tlsEnabled = $configProtocol.tls.enabled }}
-  {{- else if eq $protocol "monitor" }}
-    {{- $tlsEnabled = $.Values.config.nats.tls.enabled }}
   {{- end }}
   {{- $appProtocol := or (eq $protocol "websocket") (eq $protocol "monitor") | ternary ($tlsEnabled | ternary "https" "http") ($tlsEnabled | ternary "tls" "tcp") }}
   - {{ merge (dict "name" $protocol "targetPort" $protocol "appProtocol" $appProtocol) (omit $servicePort "enabled") (dict "port" $configProtocol.port) | toYaml | nindent 4 }}
