+++ helm/charts/nats/test/config_test.go
@@ -542,233 +542,6 @@ config:
 	RenderAndCheck(t, test, expected)
 }
 
-func TestConfigTls(t *testing.T) {
-	t.Parallel()
-	test := DefaultTest()
-	test.Values = `
-config:
-  cluster:
-    enabled: true
-    tls:
-      enabled: true
-      secretName: cluster-tls
-  nats:
-    tls:
-      enabled: true
-      secretName: nats-tls
-      merge:
-        ca_file: /etc/my-ca/ca.crt
-        verify_cert_and_check_known_urls: true
-      patch: [{op: add, path: /verify_and_map, value: true}]
-  leafnodes:
-    enabled: true
-    tls:
-      enabled: true
-      secretName: leafnodes-tls
-  websocket:
-    enabled: true
-    tls:
-      enabled: true
-      secretName: websocket-tls
-  mqtt:
-    enabled: true
-    tls:
-      enabled: true
-      secretName: mqtt-tls
-  gateway:
-    enabled: true
-    tls:
-      enabled: true
-      secretName: gateway-tls
-`
-	expected := DefaultResources(t, test)
-	expected.Conf.Value["cluster"] = map[string]any{
-		"name":         "nats",
-		"no_advertise": true,
-		"port":         int64(6222),
-		"routes": []any{
-			"tls://nats-0.nats-headless:6222",
-			"tls://nats-1.nats-headless:6222",
-			"tls://nats-2.nats-headless:6222",
-		},
-	}
-	expected.Conf.Value["leafnodes"] = map[string]any{
-		"port":         int64(7422),
-		"no_advertise": true,
-	}
-	expected.Conf.Value["websocket"] = map[string]any{
-		"port":        int64(8080),
-		"compression": true,
-	}
-	expected.Conf.Value["mqtt"] = map[string]any{
-		"port": int64(1883),
-	}
-	expected.Conf.Value["gateway"] = map[string]any{
-		"port": int64(7222),
-		"name": "nats",
-	}
-
-	replicas3 := int32(3)
-	expected.StatefulSet.Value.Spec.Replicas = &replicas3
-
-	volumes := expected.StatefulSet.Value.Spec.Template.Spec.Volumes
-	natsVm := expected.StatefulSet.Value.Spec.Template.Spec.Containers[0].VolumeMounts
-	reloaderVm := expected.StatefulSet.Value.Spec.Template.Spec.Containers[1].VolumeMounts
-	for _, protocol := range []string{"nats", "leafnodes", "websocket", "mqtt", "cluster", "gateway"} {
-		tls := map[string]any{
-			"cert_file": "/etc/nats-certs/" + protocol + "/tls.crt",
-			"key_file":  "/etc/nats-certs/" + protocol + "/tls.key",
-		}
-		if protocol == "nats" {
-			tls["ca_file"] = "/etc/my-ca/ca.crt"
-			tls["verify_cert_and_check_known_urls"] = true
-			tls["verify_and_map"] = true
-			expected.Conf.Value["tls"] = tls
-		} else {
-			expected.Conf.Value[protocol].(map[string]any)["tls"] = tls
-		}
-
-		volumes = append(volumes, corev1.Volume{
-			Name: protocol + "-tls",
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName: protocol + "-tls",
-				},
-			},
-		})
-
-		natsVm = append(natsVm, corev1.VolumeMount{
-			MountPath: "/etc/nats-certs/" + protocol,
-			Name:      protocol + "-tls",
-		})
-
-		reloaderVm = append(reloaderVm, corev1.VolumeMount{
-			MountPath: "/etc/nats-certs/" + protocol,
-			Name:      protocol + "-tls",
-		})
-	}
-
-	expected.StatefulSet.Value.Spec.Template.Spec.Volumes = volumes
-	expected.StatefulSet.Value.Spec.Template.Spec.Containers[0].VolumeMounts = natsVm
-	expected.StatefulSet.Value.Spec.Template.Spec.Containers[1].VolumeMounts = reloaderVm
-
-	// reloader certs are alphabetized
-	reloaderArgs := expected.StatefulSet.Value.Spec.Template.Spec.Containers[1].Args
-	for _, protocol := range []string{"cluster", "gateway", "leafnodes", "mqtt", "nats", "websocket"} {
-		if protocol == "nats" {
-			reloaderArgs = append(reloaderArgs, "-config", "/etc/my-ca/ca.crt")
-		}
-		reloaderArgs = append(reloaderArgs, "-config", "/etc/nats-certs/"+protocol+"/tls.crt", "-config", "/etc/nats-certs/"+protocol+"/tls.key")
-	}
-
-	expected.StatefulSet.Value.Spec.Template.Spec.Containers[1].Args = reloaderArgs
-
-	expected.StatefulSet.Value.Spec.Template.Spec.Containers[0].Ports = []corev1.ContainerPort{
-		{
-			Name:          "nats",
-			ContainerPort: 4222,
-		},
-		{
-			Name:          "leafnodes",
-			ContainerPort: 7422,
-		},
-		{
-			Name:          "websocket",
-			ContainerPort: 8080,
-		},
-		{
-			Name:          "mqtt",
-			ContainerPort: 1883,
-		},
-		{
-			Name:          "cluster",
-			ContainerPort: 6222,
-		},
-		{
-			Name:          "gateway",
-			ContainerPort: 7222,
-		},
-		{
-			Name:          "monitor",
-			ContainerPort: 8222,
-		},
-	}
-
-	expected.HeadlessService.Value.Spec.Ports = []corev1.ServicePort{
-		{
-			Name:        "nats",
-			Port:        4222,
-			TargetPort:  intstr.FromString("nats"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "leafnodes",
-			Port:        7422,
-			TargetPort:  intstr.FromString("leafnodes"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "websocket",
-			Port:        8080,
-			TargetPort:  intstr.FromString("websocket"),
-			AppProtocol: &appProtocolHTTPS,
-		},
-		{
-			Name:        "mqtt",
-			Port:        1883,
-			TargetPort:  intstr.FromString("mqtt"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "cluster",
-			Port:        6222,
-			TargetPort:  intstr.FromString("cluster"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "gateway",
-			Port:        7222,
-			TargetPort:  intstr.FromString("gateway"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "monitor",
-			Port:        8222,
-			TargetPort:  intstr.FromString("monitor"),
-			AppProtocol: &appProtocolHTTPS,
-		},
-	}
-
-	expected.Service.Value.Spec.Ports = []corev1.ServicePort{
-		{
-			Name:        "nats",
-			Port:        4222,
-			TargetPort:  intstr.FromString("nats"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "leafnodes",
-			Port:        7422,
-			TargetPort:  intstr.FromString("leafnodes"),
-			AppProtocol: &appProtocolTLS,
-		},
-		{
-			Name:        "websocket",
-			Port:        8080,
-			TargetPort:  intstr.FromString("websocket"),
-			AppProtocol: &appProtocolHTTPS,
-		},
-		{
-			Name:        "mqtt",
-			Port:        1883,
-			TargetPort:  intstr.FromString("mqtt"),
-			AppProtocol: &appProtocolTLS,
-		},
-	}
-
-	RenderAndCheck(t, test, expected)
-}
-
 func TestConfigInclude(t *testing.T) {
 	t.Parallel()
 	test := DefaultTest()
